requirements.txt
python-dotenv~=1.0.0
Flask~=2.2.3
Jinja2~=3.1.2
requests~=2.28.2